+++ doc/ref/examples/ex_objc_table_typed_intro_with_many_comments.m
@@ -1,13 +1,13 @@
 /* @@Example: not_used @@ */
-#import <Tightdb/Tightdb.h>
+#import <Realm/Realm.h>
 
 /* Defines a new table with two columns Name and Age. 
  * This table definition triggers a macro defining the classes
  * PeopleTable, PeopleTableQuery, PeopleTableRow and PeopleTableView.
- * The column types must be TightDB types (refer to the constructor 
- * TIGHTDB_TABLE_* in the documentation. */
+ * The column types must be Realm types (refer to the constructor
+ * REALM_TABLE_* in the documentation. */
 
-TIGHTDB_TABLE_2(PersonTable,    // Name of Table
+REALM_TABLE_2(PersonTable,    // Name of Table
                 Name, String,   // First column with Strings
                 Age, Int)       // Second column with Integers