+++ examples/iOSTutorial/AppDelegate.m
@@ -1,11 +1,11 @@
 #import "AppDelegate.h"
-#import <Tightdb/Tightdb.h>
+#import <Realm/Realm.h>
 
 
 // @@Example: create_table @@
 // Define table
 
-TIGHTDB_TABLE_3(PeopleTable,
+REALM_TABLE_3(PeopleTable,
                 Name, String,
                 Age,  Int,
                 Hired, Bool);
@@ -119,11 +119,11 @@ void sharedGroupFunc() {
     [fileManager removeItemAtPath:@"people.tightdb" error:&error];
 
     // @@Example: transaction @@
-    TDBContext *context = [TDBContext contextPersistedAtPath:@"people.tightdb"
+    RLMContext *context = [RLMContext contextPersistedAtPath:@"people.tightdb"
                                                             error:nil];
 
     // Start a write transaction
-    [context writeUsingBlock:^(TDBTransaction *transaction) {
+    [context writeUsingBlock:^(RLMTransaction *transaction) {
         // Get a specific table from the group
         PeopleTable *table = [transaction createTableWithName:@"employees"
                                                 asTableClass:[PeopleTable class]];
@@ -135,7 +135,7 @@ void sharedGroupFunc() {
     } error:nil];
 
     // Start a read transaction
-    [context readUsingBlock:^(TDBTransaction *transaction) {
+    [context readUsingBlock:^(RLMTransaction *transaction) {
         // Get the table
         PeopleTable *table = [transaction tableWithName:@"employees"
                                                 asTableClass:[PeopleTable class]];