+++ release_notes.md
@@ -38,11 +38,11 @@ The Objective-C API has been updated and your code will break!
 
 ### API breaking changes:
 
-* `???`
+* Changed all class prefixes from TDB to RLM
 
 ### Enhancements:
 
-* `???`
+* `createTableWithName:columns:` has been added to `TDBTransaction`.
 
 -------------